udmf : implemented higher precision vertices and thing positions.
The big change here is that the 'x' and 'y' fields have been
renamed to 'raw_x' and 'raw_y' and are generally not accessed
directly, instead they are read via new 'x()' and 'y()' methods
which convert the coordinate from fixed-point to double.

Also have SetRawX() and SetRawY() methods which not only convert
from double to fixed-point, but also round the value to integer
when the map format is not UDMF.

+++ src/r_grid.cc
@@ -128,7 +128,7 @@ int Grid_State_c::ForceSnapY(double map_y) const
 }
 
 
-int Grid_State_c::SnapX(double map_x) const
+double Grid_State_c::SnapX(double map_x) const
 {
 	if (! snap || grid.step == 0)
 		return map_x;
@@ -136,7 +136,7 @@ int Grid_State_c::SnapX(double map_x) const
 	return ForceSnapX(map_x);
 }
 
-int Grid_State_c::SnapY(double map_y) const
+double Grid_State_c::SnapY(double map_y) const
 {
 	if (! snap || grid.step == 0)
 		return map_y;
